design: Myprofile과 UserProfile 구분하여 header 구성 (#114)
